refactoring(4단계 - 로또(수동)) : defaultNumberList 수정
